refactor: rename WebsiteAppLayout to DefaultLayout and restructure layout components
- Replaced instances of WebsiteAppLayout with DefaultLayout in AppLayout.vue and updated imports accordingly.
- Removed the AppLayout and Header components, consolidating their functionality into the new DefaultLayout structure.
- Adjusted the index.ts file to reflect the new component structure and exports.
- Enhanced type definitions in index.d.ts to align with the new layout component naming.
- Cleaned up unused components related to the previous layout structure for better maintainability.
