+++ packages/utopia/src/components/layouts/DefaultLayout/components/MobileMenuButton.vue
@@ -16,7 +16,7 @@
 
 <script setup lang="ts">
 import { onMounted, onUnmounted, ref } from 'vue';
-import { Icon } from '../atoms';
+import { Icon } from '../../../atoms';
 
 interface Props {
   onOpenMenu?: () => void
