+++ packages/utopia/src/components/layouts/DefaultLayout/components/index.ts
@@ -1,0 +1,5 @@
+export { default as Header } from './Header.vue'
+export { default as Menu } from './Menu.vue'
+export { default as NavItem } from './NavItem.vue'
+export { default as MobileMenu } from './MobileMenu.vue'
+export { default as MobileMenuButton } from './MobileMenuButton.vue'
