+++ packages/utopia/src/components/layouts/DefaultLayout/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultLayout } from './DefaultLayout.vue';
